requirements.txt
dgl==1.1.1+cu118
imbalanced_learn==0.10.1
matplotlib==3.7.0
networkx==2.8.4
numpy==1.23.5
ogb==1.3.6
PyYAML==6.0
PyYAML==6.0.1
scikit_learn==1.2.1
scipy==1.10.0
torch==2.0.1
torch_geometric==2.3.1
torch_scatter==2.1.1+pt20cu118
torch_sparse==0.6.17+pt20cu118
torchmetrics==0.11.4
tqdm==4.64.1
